+++ core/contracts/tokenized-derivative/PriceFeedInterface.sol
@@ -8,7 +8,7 @@ interface PriceFeedInterface {
     /**
      * @notice An event fired when a price is published.
      */
-    event PriceUpdated(bytes32 indexed identifier, uint indexed time, int price);
+    event PriceUpdated(bytes32 indexed identifier, uint256 indexed time, int price);
 
     /**
      * @notice Whether this PriceFeeds provides prices for the given identifier.
@@ -19,5 +19,5 @@ interface PriceFeedInterface {
      * @notice Gets the latest time-price pair at which a price was published.
      * @dev Will revert if no prices have been published for this identifier.
      */
-    function latestPrice(bytes32 identifier) external view returns (uint publishTime, int price);
+    function latestPrice(bytes32 identifier) external view returns (uint256 publishTime, int price);
 }
